+++ Badge2022EF.WebApi/Models/J-Roles.cs
@@ -3,7 +3,7 @@ namespace Badge2022EF.WebApi.Models
 {
     public class J_Roles : IRoles
     {
-        public long Id { get; set; } = long.MinValue;
+        public int Id { get; set; } = int.MinValue;
         public string Name { get; set; } = string.Empty;
     }
 }
